Raspberry Pie Project
I did a secure box. This box has a lock, so to unlock it, the user has to put the correct password if not a buzzer will beep and the lock will say locked. If the user inputs the correct password, the lock will be unlocked and an email will be sent that someone unlocked this secure box.